Lol (#328) (#329)
* fix(build): re-add commands_extra + restore deleted files (#289 #292 #293)

The parallel session refactored commands.rs and lib.rs. Several of
the new modules (commands_extra, email, ftp, ai_check, observability)
that were already in HEAD got clobbered during the refactor.

- Re-add commands_extra.rs with the AppEvent IpcResponse enum using
  a non-conflicting 'event' tag (the previous 'kind' tag collided with
  the HotFolder variant field).
- Restore the missing module files (email.rs, ftp.rs, ai_check.rs,
  observability.rs) from HEAD.
- Bring lib.rs back into a state where cargo check --lib is clean.

Verified: cargo check --lib, tsc --noEmit both pass.

* chore: small follow-up tweaks from parallel sessions

- build.rs: ensure Tauri build picks up new module paths
- lib.rs: re-order mod declarations alphabetically
- tests/bleed_round_trip.rs: add fixture cleanup
- PDFView.css: harmless trailing whitespace

* feat(perf,docs,bench): #284-#298 performance + docs + benchmarks

- #284: Memoize OrderKanban (memo, useMemo, useCallback); extract
  OrderCard subcomponent so its re-render is gated on prop identity
  instead of parent re-render.
- #285: Wire vite-imagetools into vite.config.ts (PNG/JPG -> WebP/AVIF
  at build time).
- #286: react-window is already integrated through VirtualList;
  InvoiceList, ClientList, OrderListView (>=200 rows) and Dashboard
  go through it. Documented in FRONTEND-PERF.md.
- #287: Replace 'transition: all 0.2s' in OrderKanban with
  'transition: transform, opacity' + transform: translate3d(0,0,0)
  to keep the dragged card on the GPU.
- #290: cache is the hot path; documented the rationale in
  SECURITY-AUDIT.md and added db_pool.rs as a no-op shim (rusqlite
  Connection is !Sync so the lock can't be relaxed to RwLock; the
  QueryCache+Mutex<Connection> pattern in db.rs is the right shape).
- #296: validate_command_url helper (HTTPS-only, blocks loopback,
  private, link-local, CGNAT) wired into create_webhook; added
  validate_read_path for icc_profile in add_output_intent and
  check_ink_coverage path.
- #297: docs/SECURITY-AUDIT.md with full command inventory.
- #298: Criterion benches under src-tauri/benches/ for
  compress_pdf (5 MB fixture) and check_full_preflight (20-page
  fixture). Added [dev-dependencies] criterion in Cargo.toml.

New docs:
- docs/perf-baseline.md (SLOs + CI gate rationale)
- docs/FRONTEND-PERF.md (heavy components + memoization rules)
- docs/SECURITY-AUDIT.md (command surface + path/URL/SQL review)

New CI:
- .github/workflows/perf.yml (Criterion runner + PR comment)
- .github/workflows/bloat.yml (cargo-bloat top-10)
- scripts/perf.sh (local equivalent)

* chore(deps): bump lru from 0.12.5 to 0.16.3 in /src-tauri (#320)

Bumps [lru](https://github.com/jeromefroe/lru-rs) from 0.12.5 to 0.16.3.
- [Changelog](https://github.com/jeromefroe/lru-rs/blob/master/CHANGELOG.md)
- [Commits](https://github.com/jeromefroe/lru-rs/compare/0.12.5...0.16.3)

---
updated-dependencies:
- dependency-name: lru
  dependency-version: 0.16.3
  dependency-type: direct:production
...




---------

+++ src-tauri/tests/bleed_round_trip.rs
@@ -15,8 +15,8 @@ fn make_pdf_with_trim_box(trim: [f64; 4], out_path: &Path) {
             (
                 b"MediaBox".to_vec(),
                 Object::Array(vec![
-                    Object::Real(0.0),
-                    Object::Real(0.0),
+                    Object::Real(0.0 as f32),
+                    Object::Real(0.0 as f32),
                     Object::Real(trim[2] as f32),
                     Object::Real(trim[3] as f32),
                 ]),
